refactor: rename package to ovos_wolfram_alpha_plugin, migrate to pyproject.toml
- Rename ovos_wolfram_alpha_solver → ovos_wolfram_alpha_plugin
- Add pyproject.toml with full metadata and entry points
- Remove setup.py and requirements.txt
