--- ovos_wolfram_alpha_solver/version.py
@@ -1,6 +1,0 @@
-# START_VERSION_BLOCK
-VERSION_MAJOR = 0
-VERSION_MINOR = 0
-VERSION_BUILD = 5
-VERSION_ALPHA = 3
-# END_VERSION_BLOCK
