+++ pyproject.toml
@@ -1,0 +1,40 @@
+[build-system]
+requires = ["setuptools>=42", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ovos-wolfram-alpha-plugin"
+dynamic = ["version"]
+description = "ovos wolfram-alpha plugin"
+readme = "README.md"
+license = {text = "Apache-2.0"}
+authors = [
+    {name = "JarbasAi", email = "jarbasai@mailfence.com"}
+]
+keywords = ["ovos", "plugin", "wolfram"]
+dependencies = [
+    "ovos-utils>=0.7.0,<1.0.0",
+    "ovos-plugin-manager>=0.0.26,<3.0.0",
+    "ovos-config>=0.0.12,<2.0.0",
+    "requests>=2.0.0",
+    "pydantic>=2.0.0,<3.0.0"
+]
+
+[project.urls]
+Homepage = "https://github.com/OpenVoiceOS/ovos-wolfram-alpha-plugin"
+
+[project.entry-points."opm.agents.retrieval"]
+"ovos-wolfram-alpha-plugin" = "ovos_wolfram_alpha_plugin:WolframAlphaRetrievalEngine"
+
+[project.entry-points."opm.agents.toolbox"]
+"ovos-wolfram-alpha-tools" = "ovos_wolfram_alpha_plugin:WolframAlphaToolbox"
+
+[project.entry-points."opm.plugin.persona"]
+"Wolfram Alpha" = "ovos_wolfram_alpha_plugin:WOLFRAMALPHA_PERSONA"
+
+[tool.setuptools.dynamic]
+version = {attr = "ovos_wolfram_alpha_plugin.version.__version__"}
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["ovos_wolfram_alpha_plugin*"]
