--- requirements.txt
@@ -1,2 +1,0 @@
-ovos-utils>=0.7.0,<1.0.0
-ovos-plugin-manager>=0.0.26,<3.0.0
